+++ db/migrate/20130621024017_create_free_response_answers.rb
@@ -7,7 +7,7 @@ class CreateFreeResponseAnswers < ActiveRecord::Migration
       t.integer :question_id, :null => false
       t.text :free_response, :null => false, :default => ''
       t.boolean :can_be_sketched, :null => false, :default => false
-      t.decimal :credit
+      t.integer :credit
 
       t.timestamps
     end
